Moved the raycasting stuff to it's own folder
